Review fixes: - Make OGSFX test derived from GLSL one since they are nearly identical. - Add firewall check for stage source checking.
